Melhorando api

+++ src/ESX.Teste.API/Controllers/MarcaController.cs
@@ -1,4 +1,5 @@
-﻿using ESX.Teste.Application.Interfaces;
+﻿using ESX.Teste.Application.Configurations;
+using ESX.Teste.Application.Interfaces;
 using ESX.Teste.Application.ViewModels.Marca;
 using Microsoft.AspNetCore.Mvc;
 using System;
@@ -10,10 +11,13 @@ namespace ESX.Teste.API.Controllers
     public class MarcaController : ApiController
     {
         private readonly IMarcaAppService _marcaAppService;
+        private readonly IPatrimonioAppService _patrimonioAppService;
 
-        public MarcaController(IMarcaAppService marcaAppService)
+        public MarcaController(IMarcaAppService marcaAppService,
+                               IPatrimonioAppService patrimonioAppService)
         {
             _marcaAppService = marcaAppService;
+            _patrimonioAppService = patrimonioAppService;
         }
 
         [HttpGet]
@@ -26,6 +30,11 @@ namespace ESX.Teste.API.Controllers
         [Route("{id}")]
         public IActionResult GetById(Guid id)
         {
+            var patrimonio = _patrimonioAppService.GetById(id);
+
+            if (patrimonio == null)
+                return ResponseBadRequest("Marca not found");
+
             return ResponseOk(_marcaAppService.GetById(id));
         }
 
@@ -33,13 +42,7 @@ namespace ESX.Teste.API.Controllers
         [Route("{id}/patrimonios")]
         public IActionResult GetPatrimonios(Guid id)
         {
-            //RETORNAR OS PATRIMÔNIOS
-            var marca = _marcaAppService.GetById(id);
-
-            if (marca == null)
-                return NotFound();
-
-            return Ok(marca);
+            return ResponseOk(_patrimonioAppService.GetByMarcaId(id));
         }
 
         [HttpPost]
@@ -59,6 +62,9 @@ namespace ESX.Teste.API.Controllers
         [Route("{id}")]
         public IActionResult Delete(Guid id)
         {
+            if (_marcaAppService.GetById(id) == null)
+                return ResponseBadRequest("Marca not found");
+
             _marcaAppService.Remove(id);
             return ResponseOk();
         }
